Update STIMB Landing Page.py
Updated code to include root title and a new canvas to include a button to learn more. This step is still in progress as it should serve as a new page with content. Currently the additional pages are divided into separate code. Ultimately, we would like to place code on new canvases all on one code. Working progress.